+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add menu action slots to manage plugins page
+
 ## [1.0.0-alpha5] - 2023-03-14
 
 ### Added